Tempologi-master/backend/tempology/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.7
pillow==12.1.0
PyJWT==2.10.1
python-decouple==3.8
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
